api/requirements.txt
tensorflow-cpu
keras
nltk
flask
pandas
gunicorn
sklearn